Add a "private key control" check to the protocol.
Here, we require that the user also include a signature over the email address.
This serves as some proof the user controls the private key. This could be
strengthened against replay attacks by having the server issue a nonce that the user must sign.

We issue short-lived certs - another option might be to ask the user to sign a "not before" time
that is within some kind of threshold of the request.

Signed-off-by: Dan Lorenc <dlorenc@google.com>