Use 'Authorization Code with PKCE' OAuth2 flow
This change adds the use of the PKCE challenge and verifier into the
OAuth2 flow in the CLI to be compliant with RFC 8252. It also adds the
optional nonce parameter.

It also converts the Fulcio server to using the raw id_token as the
Authorization header Bearer token, and verifies that:
 - the token was signed by the specified OIDC identity provider
 - the client ID matches what is expected
 - the email address in the id_token was verified at some point in past

This change also moves the OIDC settings into command line arguments,
and shrinks the number of HTTP calls used in the OAuth2 dance.

Signed-off-by: Bob Callaway <bcallawa@redhat.com>