Launch CTFE (and deps) in docker-compose
This adds the ability to run Fulcio in docker-compose with all its
required services (a CT log which needs CTFE, trillian, and mysql). It
also ensures that on first bring up, if a trillian log has not been
created it will be and configures the CTFE instance accordingly via a
shared Docker volume.

As such, if tearing down the services using `docker-compose down`, be
sure to use the `-v` flag to ensure the shared volume is deleted as well
(otherwise a log ID will be used that no longer exists).

Signed-off-by: Bob Callaway <bcallawa@redhat.com>