Validate its a recognised CA
I had a typo in fuclio serve -ca flag and spent hours trying to work
out why the CA backend was failing. The mispelt flag was being
captured in the middleware handler instead (ca.go line 65).

This will capture it before the server even starts and make it a
lot easier to see its just a simple typo.

Signed-off-by: Luke Hinds <lhinds@redhat.com>