Add issuer from OIDC token into code signing cert
This extracts a claim from the OIDC token (defaults to `iss`) and
inserts it into the final code signing certificate issued by Fulcio.

This provides an explicit link between the identity (i.e. the email or
the URI that is the subject of the certificate) to the identity
provider who is vouching for an entity holding valid credentials for
said identity.

The value is stored in an X509 extension with OID:
1.3.6.1.4.1.57264.1.1

Signed-off-by: Bob Callaway <bob.callaway@gmail.com>