Consolidate `viper` usage in `pkg/ca/ca.go`
Using `viper` (like flags, logging) insider of libraries is a bit of a code smell, so I started to look at where we were using viper directly, and started threading particular bits of configuration through as params, which seems to be the most natural way to do this in most contexts.

Right now there is still a big wad of usage in `pkg/ca/ca.go`, which I'm going to start looking into next, but at least the remainder have been eliminated.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>