Consolidate the source-of-truth.
Today we have three copies of Certificate synthesis across:
 - x509
 - google v1
 - google v1beta1

This has burned us several times in recent memory (off the top of my head):
 - When I added Github support and missed the x509 path
 - When Scott added google v1 and dropped the K8s stuff.

By having a largely common code path for these, we should be able to avoid some of these breaks.

This change creates a method in x509ca to make an `x509.Certificate` and then reorients the bulk of the redundant certificate construction into translation logic from the shared construction.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>