Fail fast if private key is not found when using PKCS11 CA
When Fulcio is using pkcs11, it expects a private key under the
"PKCS11CA" label. If there is no such key, Fulcio starts successfully
but fails on the first signing request. A much better behavior is to
fail fast and inform the user that the private key is not found upon
start.

Closes: #284

Signed-off-by: Radoslav Gerganov <rgerganov@vmware.com>