Fix docker-compose dexidp startup (#316)
There was some change in the last year to use a non-root
user for the Dex image. This causes permission issues
when mounting files, since the non-root user cannot
access the mount. The workaround is to run as root,
which should be fine for development.

See https://github.com/dexidp/dex/issues/1649

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>