Fix null pointer crash and incorrect error statuses
The gRPC and legacy servers did not check if the public
key was provided before accesses it, causing a null pointer exception if
no public key was provided in the request.

The errors from the gRPC server to the legacy server were being wrapped,
and the original error status was being lost.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>