Add CSR support for key delivery and proof of possession
Many ecosystems and libraries have support for generating CSRs, but do
not have easy-to-use utilities for generating encoded public keys. A CSR
provides a simple way to send an encoded public key, along with a proof
of possession since CSRs are self-signed.

Existing clients do not have to change their behavior, as we will
continue to support providing a public key and signed challenge.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>