Split pkg/server from pkg/api
API clients using pkg/api.NewClient , like github.com/sigstore/cosign,
currently end up depending on all the implementation details of the servers,
unnecessarily increasing their binary size and supply-chain attack surface.

Instead, move the server code to a new pkg/server (assuming there are
no external users of that functionality), and leave pkg/api as a package only
providing a client (as it is used in practice, and documented in various
places in this repo).

The move includes pkg/api/version.go, because it is only useful
with a Go build that specifies the various values at build time,
which is unlikely to be the case for any third-party importer
of that package.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>