Add interface for certs/signer fetching to remove mutex (#643)
For all but the file CA implementation, locking when retrieving the
certs/signer is not necessary. This refactors the intermediate CA struct
by embedding an interface to fetch the certs/signer. Each CA type can
use either the basic structure that simply returns certs/signer, or one
that has a mutex to protect access to the variables.

Also small test fix - go 1.18 checks for duplicate extensions in certs
now, so I had to regenerate a test certificate.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>