Implement standardized CI extensions for GitHub
This adds the set of standardized extensions and creates the mapping for
GitHub Actions. All extension values are DER-encoded strings.

This also creates a duplicated issuer extension to match the encoding
that was used for the new extensions. OIDs 1.1 through 1.6 will be
deprecated but still present in the certificates until a future major
version of Fulcio. Updated the OID numbers so that the issuer is the
first of the new OIDs.

A future refactor will be ideal when implementing the extensions for
other CI platforms.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>