oid-info: Drop Build Signer Digest MUST -> SHOULD
It was pointed out that CI providers may not be able to meet the Build
Signer Digest requirement if they've been using server-assigned UUIDs
to identify CI runs rather than a user-verifiable digest.

Everyone seems to be in agreement that this is something CI providers SHOULD
strive for so that users have a means of detecting misbehavior by the CI
system, but this is not necessarily something we may want to gate
acceptance on, since it would require work to implement and include.

As a result, dropping this down from MUST to SHOULD.

Signed-off-by: Billy Lynch <billy@chainguard.dev>