oid-info: Remove mention of SAN from Build Signer URI.
This is not always true for all CI providers - e.g. this is true for
GitHub Actions, but for GitLab we cannot rely on it's Build Signer as a
stable identifier since it can be composed from a variety of places, so
we're planning to use the Source URI + Ref as the SAN instead.

Signed-off-by: Billy Lynch <billy@chainguard.dev>