Add persistent volume for mysql
The ctfe_init container uses a persistent volume to store the tree head
ID among other things from previous runs. However, the mysql container
is not guaranteed to reuse the same data directory when it is destroyed
and recreated, so it may come up empty. When this happens, but the CT
log is configured for an existing tree, then signing objects will fail
with a Not Found error. This change adds a persistent volume to the
docker-compose file for mysql so that the database can live on and be
consistent with the CT log's configuration.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>