ciprovider: support claim values that are not strings
The default behaviour of %v is fine in most cases:

>  bool:                    %t
>  int, int8 etc.:          %d
>  uint, uint8 etc.:        %d, %#x if printed with %#v
>  float32, complex64, etc: %g
>  string:                  %s
>  chan:                    %p
>  pointer:                 %p

However Buildkite's build_number claim is an int in the JSON, but comes
through as a Float64 and we need to render it into a string value as a
regular int.

Claim values that are floats with a fractional part will also be
converted to a string, but their fractional part will be retained. This
isn't required for Buildkite OIDC tokens, but feels like the
least-surprising behaviour for future travelers.

Signed-off-by: James Healy <james@buildkite.com>