Use Tink key APIs in `tinkca.KeyHandleToSigner`
Tink v2.3.0 [1] added the ability to extract individual keys from a
Tink `keyset.Handle` as key objects. This PR modifies
`KeyHandleToSigner` to use such APIs, significantly simplifying it.
Note that there is no need to validate keys because they are validated
when the `keyset.Handle` is created [2],[3],[4].

Also refactoring `signer_test.go` and adding a test case for an
invalid key type.

[1] https://github.com/tink-crypto/tink-go/releases/tag/v2.3.0
[2] https://github.com/tink-crypto/tink-go/blob/v2.3.0/signature/ecdsa/key.go#L380
[3] https://github.com/tink-crypto/tink-go/blob/v2.3.0/signature/ecdsa/key.go#L167
[4] https://github.com/tink-crypto/tink-go/blob/v2.3.0/signature/ed25519/key.go#L206-L209

Signed-off-by: Moreno Ambrosin <ambrosin@google.com>