fix: docker-compose.yml dependency heallthchecks and timeouts (#2013)
On slower machines, the activities in logid.sh might timeout if the mysql, log_server, and fulcio_server don't get ready in time. This causes the ct_server to constantly restart because its ct_server.cfg wouldn't be ready with an actual log id number.

This PR adjusts the dependencies and timeouts in the docker-compose.yml file.

Signed-off-by: Ramon Petgrave <ramon.petgrave64@gmail.com>