Register `/healthz` endpoint when listening on duplex http/grpc port (#2046)
* Register `/healthz` endpoint when listening on duplex http/grpc port

Registers the `/healthz` endpoint when listening on a duplex http/grpc port, just like the `http` port does when running in non-duplex mode.

The current helm charts use this endpoint for the liveness and readiness probes, so I'm assuming this should be exposed even when HTTP and gRPC are listening on the same port.

Signed-off-by: Kevin Davis <kevin.davis@chainguard.dev>

* Use _ for unused variable

Signed-off-by: Kevin Davis <kevin.davis@chainguard.dev>

---------

Signed-off-by: Kevin Davis <kevin.davis@chainguard.dev>