Handle optional environment claim (#2212)
environment is an optional claim for GitHub and GitLab tokens if
workflows/jobs don't run in an environment. Fulcio will err out if any
configured claims are missing, so as to not issue a certificate with a
missing claim unexpectedly. Because environment wasn't set, cert
issuance failed for GitHub/GitLab when this was deployed to staging.

The fix is simple thankfully: Just configure a default nil value. When
the extensions are rendered, any extensions with a empty/nil value will
be skipped over.

This PR fixes the config and adds tests to verify this fix.

This also cleans up comments in the CI provider code.

There are no functional changes in this PR, so no new release is needed.

Fixes #1845

Signed-off-by: Hayden <8418760+haydentherapper@users.noreply.github.com>
Co-authored-by: Hayden <8418760+haydentherapper@users.noreply.github.com>